Avalability Toggle

+++ src/app/middlesection/middlesection.component.ts
@@ -13,8 +13,6 @@ import { Product } from '../models/Product'
 export class MiddlesectionComponent implements OnInit {
   //instantiated array as Product array
   products:Product[] = [];
-
-
   //created a new subscription to be used when subscribing to observables
   private productSubscription: Subscription = new Subscription();
 
@@ -31,16 +29,16 @@ export class MiddlesectionComponent implements OnInit {
     .subscribe((products: Product[]) => {
       this.products = products
     })
-
-
   }
   //destroys the subscription to avoid memory leaks
   ngOnDestroy():void {
       this.productSubscription.unsubscribe();
   }
   //Deletefunction
-  onDeleteProduct(productID:string):void {
-    this.productService.deleteProduct(productID);
+  onDeleteProduct(productID:string, productName:string):void {
+    if(confirm("Do you want to remove '"+productName+"' from the inventory?")){
+      this.productService.deleteProduct(productID);
+    }
   }
   
   //handles the edit button on the top
@@ -49,11 +47,22 @@ export class MiddlesectionComponent implements OnInit {
     var element = document.querySelectorAll('td');
     editInventory?.classList.toggle('active');
     element.forEach((data)=>{
-      if(data.id == 'delete' || data.id == 'edit'){
+      if(data.id == 'delete' || data.id == 'edit' ||data.id=='availability-toggle' ||data.id=='availability-content'){
         data.classList.toggle('active');
       }
     })
     
   }
 
+  onConfirmDelete(name:string, productID:string):void{
+
+  }
+
+  toggleAvailability(product:Product):void{
+    if(product.availability == "Available"){
+      product.availability = "Not Available";
+    }else{
+      product.availability = "Available";
+    }
+  }
 }
